fix problem with SortedSet
