+++ modules/integrations/cats/src/main/scala/tethys/integration/cats/instances.scala
@@ -3,4 +3,4 @@ package tethys.integration.cats
 import tethys.integration.cats.readers.CatsReaders
 import tethys.integration.cats.writers.CatsWriters
 
-object instances extends CatsReaders with CatsWriters
+object instances extends CatsReaders with CatsWriters with NonEmptySetReader
